reynolds comments
